+++ wtforms_components/widgets.py
@@ -1,11 +1,86 @@
 from cgi import escape
 from wtforms.widgets import (
-    HTMLString, html_params, Select as BaseSelectWidget, TextInput
+    HTMLString,
+    html_params,
+    Select as _Select
 )
+from wtforms.validators import NumberRange
+from wtforms.widgets import html5
+from .validators import DateRange
 
 
-class TimeInput(TextInput):
-    input_type = 'time'
+def min_max(field, validator_class):
+    min_values = []
+    max_values = []
+    for validator in field.validators:
+        if isinstance(validator, validator_class):
+            if validator.min is not None:
+                min_values.append(validator.min)
+            if validator.max is not None:
+                max_values.append(validator.max)
+
+    data = {}
+    if min_values:
+        data['min'] = max(min_values)
+    if max_values:
+        data['max'] = min(max_values)
+    return data
+
+
+class DateTimeLocalInput(html5.DateTimeLocalInput):
+    """
+    Renders an input with type "datetime-local".
+
+    Adds min and max html5 field parameters based on field's DateRange
+    validator.
+    """
+    def __call__(self, field, **kwargs):
+        for key, value in min_max(field, DateRange).items():
+            kwargs.setdefault(key, value.strftime(field.format))
+
+        return super(DateTimeLocalInput, self).__call__(field, **kwargs)
+
+
+class DateTimeInput(html5.DateTimeInput):
+    """
+    Renders an input with type "datetime".
+
+    Adds min and max html5 field parameters based on field's DateRange
+    validator.
+    """
+    def __call__(self, field, **kwargs):
+        for key, value in min_max(field, DateRange).items():
+            kwargs.setdefault(key, value.strftime(field.format))
+
+        return super(DateTimeInput, self).__call__(field, **kwargs)
+
+
+class DateInput(html5.DateInput):
+    """
+    Renders an input with type "date".
+
+    Adds min and max html5 field parameters based on field's DateRange
+    validator.
+    """
+    def __call__(self, field, **kwargs):
+        for key, value in min_max(field, DateRange).items():
+            kwargs.setdefault(key, value.strftime(field.format))
+
+        return super(DateInput, self).__call__(field, **kwargs)
+
+
+class NumberInput(html5.NumberInput):
+    """
+    Renders an input with type "number".
+
+    Adds min and max html5 field parameters based on field's NumberRange
+    validator.
+    """
+    def __call__(self, field, **kwargs):
+        for key, value in min_max(field, NumberRange).items():
+            kwargs.setdefault(key, value)
+
+        return super(NumberInput, self).__call__(field, **kwargs)
 
 
 class ReadOnlyWidgetProxy(object):
@@ -20,7 +95,7 @@ class ReadOnlyWidgetProxy(object):
         return self.widget(field, **kwargs)
 
 
-class SelectWidget(BaseSelectWidget):
+class SelectWidget(_Select):
     """
     Add support of choices with ``optgroup`` to the ``Select`` widget.
     """